+++ app/src/main/java/org/olaven/modulist/database/dao/ModuleListDAO.kt
@@ -19,5 +19,5 @@ interface ModuleListDAO: CommonDAO<ModuleList> {
     fun getAllModuleListsLive(): LiveData<List<ModuleList>>
 
     @Query("select * from ModuleList where id = :id")
-    fun getById(id: Long): ModuleList
+    fun getById(id: Int): ModuleList
 }
