+++ app/src/main/java/org/olaven/modulist/database/model/ItemModel.kt
@@ -1,9 +1,6 @@
 package org.olaven.modulist.database.model
 
 import android.app.Application
-import android.arch.lifecycle.LiveData
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 import org.olaven.modulist.database.AppDatabase
 import org.olaven.modulist.database.entity.Item
 import org.olaven.modulist.database.repository.ItemRepository
@@ -18,7 +15,7 @@ class ItemModel(application: Application): CommonModel<Item>(application) {
 
     val allItems = repository.allItemsLive
 
-    suspend fun getByModuleListId(id: Long) =
+    suspend fun getByModuleListId(id: Int) =
         repository.getByModuleListId(id)
