+++ app/src/main/java/org/olaven/modulist/database/model/ModuleListModel.kt
@@ -2,8 +2,6 @@ package org.olaven.modulist.database.model
 
 import android.app.Application
 import android.arch.lifecycle.LiveData
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 import org.olaven.modulist.database.AppDatabase
 import org.olaven.modulist.database.entity.ModuleList
 import org.olaven.modulist.database.repository.ModuleListRepository
@@ -18,7 +16,7 @@ class ModuleListModel(application: Application): CommonModel<ModuleList>(applica
 
     val allModuleListsLive: LiveData<List<ModuleList>> = repository.allModuleListsLive
 
-    suspend fun getById(id: Long) =
+    suspend fun getById(id: Int) =
         repository.getById(id)
