+++ app/src/main/java/org/olaven/modulist/database/repository/ItemRepository.kt
@@ -10,7 +10,7 @@ class ItemRepository(private val itemDAO: ItemDAO): CommonRepository<Item>(itemD
     val allItemsLive: LiveData<List<Item>> = itemDAO.getAllItemsLive()
 
     @WorkerThread
-    suspend fun getByModuleListId(id: Long): LiveData<List<Item>> =
+    suspend fun getByModuleListId(id: Int): LiveData<List<Item>> =
             itemDAO.getByModuleListId(id)
 
 }
