+++ app/src/main/java/org/olaven/modulist/database/repository/ModuleListRepository.kt
@@ -8,6 +8,6 @@ class ModuleListRepository(private val moduleListDAO: ModuleListDAO) : CommonRep
 
     val allModuleListsLive: LiveData<List<ModuleList>> = moduleListDAO.getAllModuleListsLive()
 
-    suspend fun getById(id: Long) =
+    suspend fun getById(id: Int) =
         moduleListDAO.getById(id)
 }
